audio: auge: add HDMIRX SPDIF in support [2/2]
PD#SWPL-2956

Problem:
HDMI in DTS/Dolby input has noise when treated to LPCM

Solution:
By default,we are using PAO mode for HDMIRX,but we have not
enabled the PaPb search for 61937 raw data input, after add
that, we can detect the raw data, but it can not by clear when
switch from NONE-LPCM to NONE-LPCM, need add IRQ function to
clear that by sw.we enabled spdif in from HDMIRX, which is  the same
design as txl/txlx.we can get the Pc information now.
also we add a new interface to set spdif in source when hdmirx input.

Verify:
x301

Change-Id: I5919164e24a7e076f0e7075352b0f37fc9b4cf29
Signed-off-by: Jian Xu <jian.xu@amlogic.com>