aml_audio_hal: Improve AV sync.
PD#SWPL-27948

Problem:
When audio starts to output in tunnel mode there is
a waiting for video frame rendering ready. This was
added for Android use case and not applicable to
Linux environment and will cause underrun in continious
output mode.

Solution:
Remove the video waiting in audio output control path.

Verify:
Verified on u212 with MS12 cert test and no unerrun
at the beginning of playback.

Change-Id: Ic372cec4f5d98a81136e4c49346a824f1faba588