aml_audio_hal: Noise when ARC is disconnected with DD output. [1/1]
PD#TV-24864

[Problem]
When ARC/eARC is connected, the MS12 pipeline is
configured with a single output. If DD format is
selected (from hdmi_format), then MS12 is configured
to output DD only.
In MS12 data callback (ms12_output), ARC/eARC output
goes to hw_write for ALSA output and in alsa output
function, there is a checking to see if the data
should go to ARC/eARC to determine which ALSA device
to write. When ARC is disconnected, the flag for
the checking (bHDMIConnected) is set to 0, so the DD
data is sent to I2S alsa device, which caused noise.
DD output happens because MS12 continious output is
enabled.

[Solution]
1. In ALSA output function, when current active device
is ARC but it's disconnected, then the data will
be cleared first and still write to ARC device so
I2S ALSA device will not be injected with wrong data
prepared for ARC output.

2. When new audio stream comes, supposedly it should
trigger output reconfig and decoder reset. But the
original implementation does not have decoder reset
enabled because on customer's project ARC/eARC always
comes with a stereo PCM output, but on Linux ARC is
a single output. So a decoder reset is needed when
ARC status is changed (arc_hdmi_updated == 1)

[Verify]
1. "test_arc 1" to get ARC output in DD format
   (test_arc will force hdmi_format to 4 internally).
2. "hal_param "disconnect=262144"" to disconnect ARC
   Observe no noise from speaker.
3. test_audio_client and speaker can have normal
   single tone output.

Change-Id: Ieb3e3626ef18b586ab90de3c76daf073a8523a5e