aml_audio_hal: Fix SPDIF/HDMI output when MS12 not installed [1/1]
PD#SWPL-32639

Problem:
There is a regression for 2nd pair of PCM output when MS12 lib
is not installed. This pair of PCM output is used for SPDIF
and HDMI output, as well as speaker output on 962e2 AB311 board.

Solution:
Fix the PCM output.

Verify:
Verified on AB311 board with audio_client_test.

Change-Id: I816a780591a476ae7319779206b536ed4b07b542