audio_hal: Fix DDP output from SPDIF-A on T5 [1/1]
PD#TV-27785

Problem:
When digital output is DDP, there is a logic to
disable SPDIF output because SPDIF can only support
DD. However, on T5, SPDIF-A is hard wired to ARC output
and SPDIF-B is hard wired to SPDIF-B and SPDIF-A
->SPDIF-B same source is enabled in DTS to get
spdif output. So SPDIF output is not from SPDIF-A
but SPDIF-B. Muting SPDIF-A will mute ARC instead of
spdif output.

Solution:
Check the situation and do not mute SPDIF-A.

Verify:
Verified on T5D with ARC format set to DDP and
can get DDP from AVR.

Change-Id: I222d568a3adbc6aa25be33b44330de247ebc924f