audio_hal: Fix ARC active condition for T5/T5D [1/1]
PD#TV-28263

Problem:
For Tuner/HDMI IN, application layer only creates
audio patch with SPEAKER as destination. And when
ARC is running, if an audio patch is created with
SPEAKER spedified as output device then routing
will be changed.

Solution:
1. In fact when HDMI_ARC device connect signal is
received from set_param(), ARC/eARC is now a
"preferred" output device when available. So the
output device specified from audio patch sink
device, and stream output device are not useful.

Now totally remove the audio routing settings
from audio patch. This makes the ARC/eARC connect
and disconnect the only case to change routing.

2. In previous commit to add 2nd SPDIF output,
the pcm_open and pcm_close calling for the 2nd
SPDIF output is paired with main DIGITAL device
output in aml_audio_spdif_output.c, but there are
other places may close the main DIGITAL device
also. E.g. when audio patch is released, the close
of the main DIGITAL device pcm handle is when
continuous output status is changed, such that
we need add pcm_close in pair at those places also.
If not then when the paired pcm_open in
aml_audio_spdif_start is invoked, the 2nd
DIGITAL_SPDIF_DEVICE pcm_open will be blocked because
it was not closed earlier.

3. Fixed unnecessary eARC TX attend type checking
when eARC HW is not existing (on T5/T5D).

4. A small improvement in input threadloop. When
signal is not stable, the original in_read()
function still returns data. And depending on
different situation, pcm_read may either blocked
for 10 seconds, or returns error immediately,
which may cause either whole pipeline is underrun
or ring buffer overflow, which caused log spam
and output pcm xrun.
The new change is to return 0 when signal is not
stable, and return <0 when error and in both cases
there are no data injected to ring buffer, to avoid
unnessary data processing when data is not ready.

Verify:
Verify with T5D build.

Change-Id: I08811f4dc24b84a1705e45ccf42a85052d8deb44