audio_hal: Fix continous output flow control for patch input. [1/1]
PD#TV-29874

Problem:
When MS12 continuous output is enabled, MS12
processing is running in its own thread. There
is a rate control for the data feeding thread
to avoid too much data injected into pipeline.
When continuous output was enabled in previous
CL, this rate control is also applied to patch
case.

The rate control has a problem for IEC61937
format, which is the main format for HDMI IN
non-pcm input. The time calculation for IEC61937
is not there so the rate control through virtual
buffer is not working correctly for IEC61937 format.

And for patch input case, it's not needed to do
rate control anyway.

Solution:
Disable the use of rate control for audio patch case.

Verify:
Verify with DDP input from HDMI IN. Switch in
and out with HDMI IN for many times and observe
audio output is always smooth.

Change-Id: Ia3f050438d1898b6d1ff71cc5807bdbc4da460e9