aml_audio_hal: Fix master volume settings. [1/1]
PD#SWPL-43192

Problem:
Volume settings with MS12 are applied at two places.
For Dolby format and Direct PCM input, which are
the most common cast with RDK GST pipeline, the volume
settings are implemented by MS12 main volume control.
Another place to apply the volume change is in the
data_processing function for MS12 processed PCM output.

The master volume setting is by apply the volume to
all the output streams for all use cases. When MS12
works in continuous mode, a separate MS12 output
stream is created by copy the original output stream's
settings, including the volume settings. Therefore the
initial stream volume is saved to MS12's continuous
stream's volume and also being applied to MS12's main
volume control.

There are two issues:
1. For PCM output from MS12, the volume gain is applied
twice. One in MS12 and the other in data_processing after
MS12. So the output volume is not correct.

2. If a volume is set before injecting any data, the volume
is carried to MS12 continuous output stream and used in
data_processing. And the following volume setting only change
the original stream's volume and MS12's main volume and the
volume for MS12 continuous stream's volume is kept at
the initial value when it's constructed. Considering this
sequence:

  1) Set master volume to 0 or set master_mute to 1
  2) Open a new output stream and because of global master
     mute, the stream's volume is set to 0.
  3) Write stream data and trigger MS12 pipeline created
     with a new MS12 stream constructed with same volume
     as 0.
  4) Change master volume. The master volume affects the
     original stream's volume and also applied to MS12's
     main volume gain.
  5) In data_processing, MS12's volume gain 0 is applied
     again, causing no audio output.

Solution:
For MS12 output, the volume gain in data_processing should
not be applied since they are controlled by MS12 main volume
already.

Verify:
Verified with the sequence on DSTV platform with the
operating sequence.

Signed-off-by: Tim Yao <tim.yao@amlogic.com>
Change-Id: I1eeaa8a95f9ae0b8d72f9e80d742b2de3df71dc1