audio_hal: Fix ms12_bypass memory leak for PCM only output. [1/1]
PD#TV-31581

Problem:
In order to support MS12 bypass output there is a link
list created for all the AC3/EAC3 frame data. The data
are used to replace the output from MS12's output when
bypass mode is used.
The link list works as a buffer pool and the data are
injected at same time when data are sent to MS12 pipeline.
The comsuming of this data is from MS12 output callback
when compressed format data is processed.
When MS12 is configured with PCM only, e.g. when the
device is a TV with only stereo downmix and DAP speaker
output configured, there is no compressed output from
MS12 then this buffer pool has no consumer, causing
memory consumption getting bigger and bigger during
the playback.

Solution:
When there is no compressed output from MS12 configured,
the compressed data in bypass pool are also consumed.

Verify:
Test with a modified version of halplay to loop play
ac3 source file and use top command to verify memory
comsumption of audio_server process to make sure the
memory consumption is not growing over the time.
The test must be performed with long duration to make
sure there is no other leaks during playback.

Change-Id: I55e3abdfd68ba59a824ae091fbba3d7e877492b4
Signed-off-by: Tim Yao <tim.yao@amlogic.com>