audio_hal: Add environmental setting support [1/1]
PD#

Problem:
Android properties are implemented as Linux env
settings but when audio server is running the
env variable changes are not reflected in a running
process. So the debug dump or tuning of parameters
requires a restart of audio_server.

Solution:
Add a new "setenv" device set_parameter command
so hal_param can be used to set/unset env variables
within audio_server process.

Verify:
Verified with AB301 build.

Change-Id: I88a24f0d5d78c4b4c9e33c32d9e66512f8e372a3