audio_hal: Add Netflix PTS gap processing. [1/1]
PD#SWPL-45351

Problem:
Netflix NRDP may send EAC3 audio stream down with
PTS gaps. The requirement is to fill the gap with
silence and also do volume ramping before and after
the silence.

Solution:
1. Add the set_param interface to accept gap start
point (in byte offset of the stream) and the duration
of silence. The forward the information to MS12 lib.
The volume ramping and silence will be handled inside
MS12.

2. Because of PTS gap, the APTS look up returned
from MS12 can have a jump because of inserted silence
inside MS12 at the insert point. But it will be recover
after silence period. Such PTS jump may cause audio
data insert as a result of AV sync because audio PTS
is ahead of PCR (which is a fake report). To solve
this problem, a flag is added to skip AV sync action
during this slience period.

Requires MS12 lib update:
https://scgit.amlogic.com/#/c/152594/

Verify:
Verify with NTS PLIST test with waveform dump.

Change-Id: Iaafaa0fb3cef0ee54f6c425b8d8715c8fe289d58
Signed-off-by: Tim Yao <tim.yao@amlogic.com>