audiohal: Fix ARC/eARC with eARC path for HDMI INput under passthrough mode. [1/1]
PD#TV-33877

Problem:
For HDMI IN when input is bit stream with SPDIF 61937
encoded format, the code path uses bypass_process when
mode is set to bypass for HDMI IN passthrough.
The changes made for ARC/eARC output through eARC path
does not exist for the case so there is no output.

Solution:
Add the same processing for eARC path in same way as
MS12 output for bit stream output and use hw_write for
ARC output.

Verify:
Verified on HDMI IN source with DD/DDP format and
output mode set to bypass.

Change-Id: I6517d681ff86e12eef1a5bb455733f43c4f34fcb
Signed-off-by: Tim Yao <tim.yao@amlogic.com>