audiohal: Refine pcr master msync processing. [1/1]
PD#SWPL-37180

Problem:
When first audio data injected and msync control
returns ASYNC_AGAIN the data is dropped but the first
PTS of dropped data is added into PTS table for lookup.
And when future PTS records are added, more PTS entries
with offset 0 are added into the table. The implementation
of PTS lookup uses first PTS record which caused the first
dropped audio PTS is always used. And then the msync
AV sync is broken because the increment of first start
APTS is not reflected.

Solution:
When offset is same, the new check in PTS record
overwrite the same entry so the new APTS will be
used.

Verify:
Verify with the case when msync start audio return
ASYNC_AGAIN to drop data.

Change-Id: If5f5733a2f707892d71649aeac224042dd49868f