audiohal: Fix ARC/eARC passthrough output. [1/1]
PD#TV-35456

Problem:
For eARC output in passthrough output, the passthrough
checkin data is not used and ms12 encoded output is
used.

Solution:
Use the checked in passthrough data.

Also whe passthrough is specified, remote additional
max audio cap checking because in real project for
ARC connection the audio cap for ARC is not received
whe passthrough mode is selected. If the checking
is applied, then AVR may only see PCM output when
passthrough mode is used, which is confusing. So
we select to ignore the checking, even it may cause
no output if sink side does not support the format.

Verify:
Verified with MS12 cert test 1285 with 7.1 as source
and we should see 7.1 output from ARC output when
passthrough mode is used.

Change-Id: I521fff1c5b2aac8e61339f150fb8af8f24ef3e5d