audio_hal: add MS12 master volume support   [1/1]
PD#OTT-24700

Problem:
When the box digtal_mod selects 5, and the connected TV
has strong support capabilities, for example, it can
support Dolby format,customer plays multimedia files,
the HAL API cannot be used to adjust the global volume,
because the output supports strong format, after encoding,
the MS12 volume adjustment interface is not added
in adev_set_master_vloume.

Solution:
1.When master volume is set, the streams for each active
stream will be set with an additional master volume
gain applied.
2.For the case of ms12 lib, use set_ms12_main_volume
to adjust the volume.
3.Because the volume is set through the stream, in order
to avoid setting the master volume when there is no
stream, the next time the playback volume is invalid,
when the stream is is opened, call out_set_volume() to
set the volume.
4.Re-modify the mute interface.
5.sync to the solution of the 12063 problem in Gerrit.

Verify:
Verify OTT-24700 issue

Change-Id: Ib60fd6bd7d3fb0cd03eceae9b30bc11863759398
Signed-off-by: hui.liu <hui.liu@amlogic.com>