audiohal: Do not bypass MS12 for netflix system sound [1/1]
PD#SWPL-107739

Problem:
On NRDP6 the Netflix UI sound is 6 channel PCM, which
is injected to audio hal as system sound input. There
is a condition to bypass MS12 for multichannel and hi-res
PCM input in audio hal, and then use a broken SW mixer
path to handle the input stream, and that path was designed
for main input only since the system sound is assumed
to be 2 channel only.

Solution:
With 5.1 channel system sound support already added,
it's not necessary to bypass MS12 for multi-channel
system sound in passthrough mode. So exclude Netflix
multichannel PCM system sound as a condition to bypass
MS12.

Verify:
NTS AUDIO-007-DDP51-PASSTHROUGH

Change-Id: Idb1aaef933de1f7557951a6b973953f7c28c9d94
Signed-off-by: Tim Yao <tim.yao@amlogic.com>