audio: Fix pop noise when play audio only file [1/1]
PD#SWPL-136300

Problem:
When av_sync_audio_render is not done, policy.action is not updated and
uses the default value of rendor, which eventually causes policy.action
to frequently switch between rendor and insert, and then causes a pop sound.

Solution:
1.Initialize policy.action to the last updated value instead of the default
value, and policy.action will not switch between the default value and the
updated value;
2.This is a supplementary modification to cl:
https://scgit.amlogic.com/#/c/341223.

Verify:
yocto_ap222.

Change-Id: I35dfe2eb78dce6f87e69cc2d1c9eaa01169d2373
Signed-off-by: yuliang.hu <yuliang.hu@amlogic.com>